Developer Documentation

API

Main.ConwaysGameOfLife.get_living_neighbours — Method

getlivingneighbours(cell) Returns the number of living cells in the neighbours Array of 'cell'

Arguments

cell::Cell: The cell to be checked

Examples

julia> cell = Cell()
julia> cell1 = Cell()
julia> set_alive(cell1)
julia> cell2 = Cell()
julia> set_dead(cell2)
julia> set_neighbours(cell,cell1)
julia> set_neighbours(cell,cell2)
julia> get_living_neighbours(cell)
1

Main.ConwaysGameOfLife.get_option — Method

getoption(option) Returns different matrixes to use as an initstate for Conways Game of Life depending on 'option'. May require a readline() to determine the size.

Arguments

option::String: String to get the corresponding option

Examples

julia> option = get_option("Random")
julia> option
rand([0,1], y_size, x_size)

Main.ConwaysGameOfLife.main — Method

main() Initializes and runs a game of Conways Game of Life. It also initializes the interactive function of the GLMakie Buttons.

Main.ConwaysGameOfLife.pulsar — Method

spacecannon() Returns a matrix of a pulsar to use as an initstate for Conways Game of Life

Main.ConwaysGameOfLife.set_alive — Method

set_dead(cell) Sets the alive attribute of 'cell' to false

Arguments

cell::Cell: The cell to be set dead

Examples

julia> cell = Cell()
julia> set_dead(cell)
julia> cell.alive
false

Main.ConwaysGameOfLife.set_button — Method

set_button(cell, button) Sets the button attribute of 'cell' to 'button'. The button is from the GLMakie package

Arguments

cell::Cell: The cell to change the button

button::Button: The button to be set to the cell

Examples

julia> cell = Cell()
julia> fig = Figure()
julia> green_button = Button(fig[1,1], buttoncolor = green)
julia> set_button(cell,button)
julia> cell.button
green_button

Main.ConwaysGameOfLife.set_dead — Method

set_alive(cell) Sets the alive attribute of 'cell' to true

Arguments

cell::Cell: The cell to be set alive

Examples

julia> cell = Cell()
julia> set_alive(cell)
julia> cell.alive
true

Main.ConwaysGameOfLife.set_neighbours — Method

set_neighbours(cell, neighbours) Appends a 'neighbours' to the neighbours array of 'cell'

Arguments

cell::Cell: The cell to change the button

neighbours::Cell: The cell to be added to neighbours

Examples

julia> cell = Cell()
julia> cell_new = Cell()
julia> set_neighbours(cell,cell_new)
julia> cell.neighbours
Array([cell_new])

Main.ConwaysGameOfLife.should_update — Method

shouldupdatecell(cell) Checks whether 'cell' should be updated for the next iteration of Conways Game of Life

Arguments

cell::Cell: The cell to be checked

Examples

julia> should_update_cell(living_cell_with_four_living_neighbours)
false

julia> should_update_cell(dead_cell_with_three_living_neighbours)
true

julia> should_update_cell(living_cell_with_two_living_neighbour)
false

Main.ConwaysGameOfLife.should_update_cell_alive — Method

shouldupdatecell_alive(cell) Checks whether 'cell' should be updated for the next iteration of Conways Game of Life assuming 'cell.alive = true'

Arguments

cell::Cell: The cell to be checked

Examples

julia> should_update_cell_alive(cell_with_four_living_neighbours)
true

julia> should_update_cell_alive(cell_with_three_living_neighbours)
false

julia> should_update_cell_alive(cell_with_one_living_neighbour)
true

Main.ConwaysGameOfLife.should_update_cell_dead — Method

shouldupdatecell_dead(cell) Checks whether 'cell' should be updated for the next iteration of Conways Game of Life assuming 'cell.alive = false'

Arguments

cell::Cell: The cell to be checked

Examples

julia> should_update_cell_dead(cell_with_four_living_neighbours)
false

julia> should_update_cell_dead(cell_with_three_living_neighbours)
true

julia> should_update_cell_dead(cell_with_two_living_neighbour)
false

Main.ConwaysGameOfLife.space_cannon — Method

spacecannon() Returns a matrix of a glider cannon to use as an initstate for Conways Game of Life

Main.ConwaysGameOfLife.update_cell — Method

update_cell(cell) Updates the 'color' and 'alive' attribute of 'cell' if the 'update' attribute is true. White, dead cells will become black, alive cells and vice versa.

Arguments

cell::Cell: The cell to be updated

Examples

julia> update_cell(white_dead_cell_update_true)
true

julia> update_cell(white_dead_cell_update_false)
false

julia> update_cell(black_alive_cell_update_true)
true

Main.ConwaysGameOfLife.update_game — Method

update_game(gameboard) initializes the next iteration of Conways Game of Life

Arguments

gameboard::Gameboard: The current iteration of the gameboard

Main.ConwaysGameOfLife.Cell — Type

This mutable struct initializes and contains all the information needed to simulate a single cell in Conways Game of Life

Fields

'neighbours::Array': An array with all neighbours of the cell

'alive::Bool': A bool stating whether the cell is alive or note

'update::Bool': A bool stating wheter the cell should be updated in the next iteration

'button::Button': A GLMakie button to visualize the cell

Main.ConwaysGameOfLife.Gameboard — Type

This mutable struct initializes and contains all the information needed to simulate Conways Game of Life

Fields

'initial_state::Matrix': Matrix filled with 0's and 1's. 0's are initialized as dead cells and 1's as living cells on their respective position in the matrix.

'currentstate::Matrix': A matrix of the currentiteration of the game. Every entry is of type 'Cell' and contains the relevant information for the next iteration.

'gamesize::Tuple': A tuple with the dimensions of 'initialstate' in form of (y,x)

'fig::Figure': A GLMakie 'Figure' to visualize the game

'nextButton::Button': A GLMakie 'Button' to initialize the next iteration

Index

Main.ConwaysGameOfLife.Cell

Main.ConwaysGameOfLife.Gameboard

Main.ConwaysGameOfLife.get_living_neighbours

Main.ConwaysGameOfLife.get_option

Main.ConwaysGameOfLife.main

Main.ConwaysGameOfLife.pulsar

Main.ConwaysGameOfLife.set_alive

Main.ConwaysGameOfLife.set_button

Main.ConwaysGameOfLife.set_dead

Main.ConwaysGameOfLife.set_neighbours

Main.ConwaysGameOfLife.should_update

Main.ConwaysGameOfLife.should_update_cell_alive

Main.ConwaysGameOfLife.should_update_cell_dead

Main.ConwaysGameOfLife.space_cannon

Main.ConwaysGameOfLife.update_cell

Main.ConwaysGameOfLife.update_game